refactor: convert ping API to async/await
