--- src/core/components/ping-readable-stream.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const toStream = require('pull-stream-to-stream')
-
-module.exports = function pingReadableStream (self) {
-  return (peerId, opts) => toStream.source(self.pingPullStream(peerId, opts))
-}
